alert-service/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.2
redis==5.0.7